多线程已经改好，流的临界区问题尚未解决，开始调试

+++ core/parser.h
@@ -15,7 +15,7 @@ typedef struct packet {
 #define PKT_DIR_MASK 0x3
 #define PKT_DEBUG 0x1000000000000000
     uint64_t flag:62;/*其他标志*/
-	uint64_t pktag;
+	uint32_t pktag;
 	uint8_t prot_types[MAX_PARSER_DEPTH];
     uint8_t prot_offsets[MAX_PARSER_DEPTH];
     uint8_t prot_depth;
